Added Result to takeTurn, about to create InteractivePlayer
